Create new component CopyToClipboard and use it in create project table flow
